Updated logging

+++ src/commands/DeleteQuestCommand.js
@@ -3,9 +3,9 @@
 import Logger from '../utils/Logger';
 
 export default class {
-  constructor (slack, questsService) {
+  constructor(slack, questsService) {
     this.questsService = questsService;
-    this.logger = new Logger('CreateQuestCommand');
+    this.logger = new Logger('DeleteQuestCommand');
 
     slack.on('/quest-delete', async (msg, bot) => {
       this.logger.log('Processing message', msg.text);
@@ -15,8 +15,6 @@ export default class {
         return bot.replyPrivate('Invalid questId');
       }
 
-      this.logger.log('Matches', matches);
-
       const questId = matches[0];
 
       try {
@@ -24,13 +22,19 @@ export default class {
         return bot.replyPrivate('Quest deleted!');
       } catch (e) {
         this.logger.error(e);
-        return bot.replyPrivate('Whoops! There\'s been an error!');
+        return bot.replyPrivate(e.message);
       }
     });
 
   }
 
-  execute (questId) {
+  async execute(questId) {
+    const quest = await this.questsService.get(questId);
+
+    if (!quest) {
+      throw new Error('Invalid questId');
+    }
+
     return this.questsService.delete(questId);
   }
 }
